prior_proc_ALL.r - get subset of clairNBS_Prior
At line 96, clairNBS was use to get the subset of ClairNBS_Prior rather than the ClairNBS_Prior. As clairNBS object has been restructured (potentially with less columns if some of observed data was turn off in config.csv file). 

This inconsistency can lead to incorrect column was used as residNBS and potentially crash the program in some instances.